Move decorators to a new module
