+++ tests/test_cli.py
@@ -1,7 +1,6 @@
 """Basic tests tha cli app is running"""
 import os
 from pizza.pizza_menu import pizza_menu
-import pizza.business  # to disable latency by importing before cli
 from pizza.cli import cli
 from click.testing import CliRunner
 from .help_funcs import all_pizzas_parameters, all_types_delivery
